requirements.txt
requests>=2.28.0
streamlit>=1.32.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.0
openai>=1.13.0
PyPDF2>=3.0.1
termcolor>=2.4.0
tqdm>=4.66.0
validators>=0.22.0